46.Authorizing action

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Models\Task;
+use Illuminate\Http\Request;
 
 class TaskController extends Controller
 {
@@ -14,8 +15,11 @@ class TaskController extends Controller
      */
     public function index()
     {
-        // return TaskResource::collection(Task::all());
-        // return Task::all()->toResourceCollection();
+
+        if ($request->user()->cannot('viewAny', Task::class)) {
+            abort(403);
+        }
+
         return request()->user->tasks()->get();
     }
 
@@ -24,6 +28,11 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+
+        if ($request->user()->cannot('create', Task::class)) {
+            abort(403);
+        }
+
         // $task = Task::create($request->validated()+['user_id'=>$request->user()->id]);
         $task = $request->user()->tasks()->create($request->validate());
 
@@ -33,10 +42,11 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Task $task)
+    public function show(Task $task, Request $request)
     {
-        // return new TaskResource($task);
-        // return TaskResource::make($task);
+        if ($request->user()->cannot('view', $task)) {
+            abort(403);
+        }
         return $task->toResource();
     }
 
@@ -46,6 +56,9 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if ($request->user()->cannot('update', $task)) {
+            abort(403);
+        }
         $task->update($request->validated());
 
         return $task->toResource();
@@ -54,8 +67,12 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task, Request $request)
     {
+        if ($request->user()->cannot('delete', $task)) {
+            abort(403);
+        }
+
         $task->delete();
 
         return response()->noContent();
